Added popup message to shipwreck

+++ Assets/Scripts/SubmarineComponents/Battery.cs
@@ -1,8 +1,8 @@
 ﻿public class Battery
 {
-    public int AmountToCharge;
+    public float AmountToCharge;
     
-    public Battery(int amountToCharge)
+    public Battery(float amountToCharge)
     {
         AmountToCharge = amountToCharge;
     }
